wip sega_screen [+] feat: megadrive color type [+] feat: fadeout functions [+] feat: API for vpu, plc, interrupt [+] feat: s1disasm ResourceIDs by Adol [+] tests: tinyint sizes [+] feat: test framework two values equal [~] refactor: little project structure refactor

+++ include_backend/tinyint.h
@@ -3,12 +3,12 @@
 
 typedef     signed char         i8;
 typedef     signed short        i16;
-typedef     signed long         i32;
+typedef     signed int          i32;
 typedef     signed long long    i64;
 
 typedef     unsigned char       u8;
 typedef     unsigned short      u16;
-typedef     unsigned long       u32;
+typedef     unsigned int        u32;
 typedef     unsigned long long  u64;
 
 typedef     unsigned long long  size;
